--- source/license/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class LicenseConfig(AppConfig):
-    name = 'license'
